server: make status checking case-insensitive
This PR makes string comparisons for `Status` and `Resolution` use
`strings.EqualFold` instead of `==`. Jira statuses are case-insensitive.
They are usually displayed in full caps via the UI, and that is what we
use for the config and when updating the status. However, the status and
resolution returned via `GetIssue` only capitalize the first letter.