server: use blocks, not clones, for cherrypicks
This PR changes the cherrypick process to also add the `Blocks` link
type to dependent bugs and only use `Blocks` and `Depend` type links to
identify dependent bugs. It also updates the backport instruction
message to match the latest changes to the backporting FAQ.

Further, it cleans up some fields that were unnecessarily included in
some of the `validateBug` tests.