go.mod: update vulnerable dependencies
This updates the `yaml.v3` dependency to `v3.0.1` by removing the
`replace` that had been copied from `test-infra` as a vulnerable
function was being used. This also updates `golang.org/x/net` to
`v0.4.0`. There was a vulnerability in the previous version, but we did
not execute the vulnerable functions, so it was not an urgent fix.