server: allow multiple bugs per PR
This PR adds the ability for users to add multiple bugs per PR. It will
make a single comment with the usual message repeated multiple times for
the multiple PRs, separated by 2 newlines. For retitled PRs
(cherrypicks) the retitle will be moved after all of the messages in
order to properly update all bugs in a title.

If multiple bugs have different severity levels, the highest priority
severity level will be used for the PR label.

For errors, the plugin will return just the error for standard `handle`
function errors, as `handle` can be repeated by running `/jira refresh`.
For `close`, `merge`, and `cherrypick` events, which cannot be repeated,
the plugin will continue to handle later bugs and make a full comment to
indicate the operations were performed on all bugs instead of just
reporting on the failure.