server: don't make duplicate comments
This PR adds a check to prevent the bot from making identical comments
to a PR when nothing relevant has changed, a manual refresh wasn't asked
for, and an identical comment was already made by the bot. This should
prevent the bot from spamming comments in certain situations, such as
checkboxes being updated in a PR's description.