server,helpers: add support for update assignee and Sprint during clone
This PR adds the `Assignee` as one of the values that gets updated after
bug creation during the cherrypick cloning process. It also adds basic
support for cloning the `Sprint` value during the post-clone update, but
does not add the struct as the plugin does not need to parse any data
from it. While `OCPBUGS` does not use the `Sprint` field, other projects
do and while the plugin does not fully support non-OCPBUGS projects,
this did not require too much extra effort to implement.