server: wait for jira assignee automation
This PR fixes a race condition where the inbuilt jira automation sets
the Assignee of the clone to the project default after the plugin
updates the Assignee to match the parent. This PR also moves the sprint
field update into the same operation as the Target Version and Assignee
update.