server: add backport command
Adds a `/jira backport` command that takes a list of comma-separated
branches (similar to `/jira cherrypick`'s issue keys) and creates the
corresponding issues for the branches with the correct `Blocks` and
`Clones` types for backporting. The command will first verify that a
continuous link between all issues and the original parent can be
created before starting the cloning process.

Once all clones are created and linked, the original parent issue will
be updated with new labels that will allow the plugin to identify the
correct bug to link when the new cherrypicks are created. It will then
create a comment queuing up cherrypicks for all specified branches.