.devcontainer: add devcontainer
This PR adds an initial devcontainer json and Containerfile that uses the
openshift-4.17 golang-1.22 builder image, makes adjustments to allow it to run
in rootless podman, and installs commonly used golang tools. It also configures
the vscode to enable the go and docker extensions in the devcontainer. Note
that users will need to be logged into the quay-proxy registry to pull the base
image for the Containerfile.