server: fix backport bug caused by z-stream versions
Fix a bug that was causing the plugin to fail to correctly create the
childBranches map when a branch has both `.0` and `.z` dependents, but only a
branch for the `.z` target version exists. Also handle situation where one
target version could be required by multiple branches (for example, `main` and
`release-4.X`).