OCPCRT-385: server: add new premerge verification method (#124)
* OCPCRT-385: server: add new premerge verification method

Add new premerge verification method using the `/verified by` and `/verified
later` commands.  When a user sends a `/verified by` or `/verified later`
comment, the program will check that it is a valid request. If the request is
valid, it will then add either a `verified` or `verified-later` label to the PR
and upload a struct to bigquery to allow future data analysis. If a PR is
merged and has the `verified` label applied, it will bypass the usual on merge
status update defined in the config file and set the status to `VERIFIED`
instead.

* Add timestamp to bigquery data and tweak verified later comment