server: allow spaces in verification reasons
Update regex to allow spaces in verification reasons due to some tests have
spaces in their names. This also adds trimming of leading and trailing spaces
in the reasons and simplifies the regexes for verified, verified later, and
backport, as the comma part of the regex was unnecessary since we just do a
string split on the comma.