Allow multiple commands
It's not uncommon to have multiple commands in a comment. For example:

  /jira refresh
  /label backport-risk-assessed
  /verified by @stephenfin

This doesn't currently work since we insist the command is left in a
unique comment. This generates noise and is confusing. Fix the issue by
allowing users to combine commands with other commands. This also allows
for multiple commands to be issued to this plugin at the same time.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>